==> 05. Imdb Score Model/01. Create Dataset.ipynb <==
# https://www.imdb.com/interfaces/
# https://datasets.imdbws.com

# title.basics 
#   ID Film / Movie / Title / Year / Genres 

# title.ratings
#   ID Film / Note 

# title.crew
#   ID Film / ID Director 

# name.basics
#   ID Director / Name 

import pandas as pd
import numpy as np

title_basics = pd.read_table("src/title.basics.tsv", on_bad_lines='skip', low_memory=False)
title_basics = title_basics[title_basics['titleType'] == 'movie']
title_basics = title_basics.drop(columns=['titleType','primaryTitle','isAdult', 'endYear','runtimeMinutes'])
title_basics = title_basics[title_basics['originalTitle'] != r'\N']
title_basics = title_basics[title_basics['startYear'] != r'\N']
title_basics = title_basics[title_basics['genres'] != r'\N']

title_ratings = pd.read_table("src/title.ratings.tsv", on_bad_lines='skip', low_memory=False)
title_basics_ratings = pd.merge(title_basics, title_ratings, left_on='tconst', right_on='tconst')

title_crew = pd.read_table("src/title.crew.tsv", on_bad_lines='skip', low_memory=False)
title_crew = title_crew.drop(columns=['writers'])
title_basics_ratings_crew = pd.merge(title_basics_ratings, title_crew, left_on='tconst', right_on='tconst')
title_basics_ratings_crew["directors"] = [ str(x).replace(' ', '') for x in title_basics_ratings_crew["directors"][:] ] 
title_basics_ratings_crew["directors"] = [ x.split(",") for x in title_basics_ratings_crew["directors"][:] ]
title_basics_ratings_crew = title_basics_ratings_crew.explode('directors').reset_index()

name_basics = pd.read_table("src/name.basics.tsv", on_bad_lines='skip', low_memory=False)
name_basics = name_basics.drop(columns=['birthYear','deathYear','primaryProfession','knownForTitles'])
title_basics_ratings_crew_name = pd.merge(title_basics_ratings_crew, name_basics, left_on='directors', right_on='nconst')
title_basics_ratings_crew_name = title_basics_ratings_crew_name.drop(columns=['index','directors','nconst'])
title_basics_ratings_crew_name["startYear"] = pd.to_numeric(title_basics_ratings_crew_name["startYear"]).astype(int)

import urllib.parse
# https://www.imdb.com/search/title/?title=TOUT+S%27EST+BIEN+PASSE&release_date=2015-01-01,2022-12-31

url_base = 'https://www.imdb.com/search/title/?title='

title_basics_ratings_crew_name['imdb_search'] = ''

for i in range( len(title_basics_ratings_crew_name) ):

    title = urllib.parse.quote(title_basics_ratings_crew_name.iloc[i]['originalTitle']).replace('%20', '+')
    date = str(title_basics_ratings_crew_name.iloc[i]['startYear'])

    title_basics_ratings_crew_name.imdb_search[i] = f"{url_base}{title}&release_date={date}-01-01,{date}-12-31"
    
title_basics_ratings_crew_name.head()

title_basics_ratings_crew_name.to_csv(r'export/dataset_without_resume.csv', index = False, header=True)
