==> 05. Imdb Score Model/03.Bis Cleaning Dataset.ipynb <==
# https://www.youtube.com/watch?app=desktop&v=IEEhzQoKtQU&feature=youtu.be
# https://www.sbert.net/docs/quickstart.html#comparing-sentence-similarities

import pandas as pd
import numpy as np
import threading
from sentence_transformers import SentenceTransformer, util
import warnings
warnings.filterwarnings("ignore")

dataset = pd.read_csv("export/temps/dataset.csv", delimiter=',', on_bad_lines='skip')
model = SentenceTransformer('all-MiniLM-L6-v2')
print(len(dataset))
dataset = dataset.drop_duplicates()
print(len(dataset))

# Convert string Genre to list
dataset = dataset.rename(columns={"genres": "genre"})
dataset["genre"] = [ str(x).replace(' ', '') for x in dataset["genre"][:] ]
dataset["genres"] = [ x.split(",") for x in dataset["genre"][:] ]

model = SentenceTransformer('all-MiniLM-L6-v2')

dataset['genre'] = ''
dataset['themaScore'] = 0.0

def compute_themaScore(dataset):
    data = dataset.copy()

    for i in range( len(data) ):
        bestScore = 0.0
        bestGenre = ''
        for j in range(len(data.iloc[i]['genres'])):
            result = util.cos_sim(model.encode(str(data.iloc[i]['genres'][j])), model.encode(str(data.iloc[i]['resume']))).tolist()[0][0]
            if result > bestScore:
                bestScore = result
                bestGenre = data.iloc[i]['genres'][j]

        #data.loc[i, 'themaScore'] = bestScore
        #data.loc[i, 'genre'] = bestGenre
        data['themaScore'][i] = bestScore
        data['genre'][i] = bestGenre

        #print(i,'/',len(data),bestGenre, bestScore)
        #print(data.iloc[i])

    return data

t11 = compute_themaScore(dataset[290000:])
t11 = t11.drop(columns=['genres'])
t11.to_csv(r'temps/t11.csv', index = False, header=True)



#t1 = compute_themaScore(dataset[:30000])
#t1 = t1.drop(columns=['genres'])
#t1.to_csv(r'temps/t1.csv', index = False, header=True)
'''
t2 = compute_themaScore(dataset[30000:60000])
t2 = t2.drop(columns=['genres'])
t2.to_csv(r'temps/t2.csv', index = False, header=True)

t3 = compute_themaScore(dataset[60000:90000])
t3 = t3.drop(columns=['genres'])
t3.to_csv(r'temps/t3.csv', index = False, header=True)

t4 = compute_themaScore(dataset[90000:120000])
t4 = t4.drop(columns=['genres'])
t4.to_csv(r'temps/t4.csv', index = False, header=True)

t5 = compute_themaScore(dataset[120000:150000])
t5 = t5.drop(columns=['genres'])
t5.to_csv(r'temps/t5.csv', index = False, header=True)

t6 = compute_themaScore(dataset[150000:180000])
t6 = t6.drop(columns=['genres'])
t6.to_csv(r'temps/t6.csv', index = False, header=True)

t7 = compute_themaScore(dataset[180000:210000])
t7 = t7.drop(columns=['genres'])
t7.to_csv(r'temps/t7.csv', index = False, header=True)

t8 = compute_themaScore(dataset[210000:240000])
t8 = t8.drop(columns=['genres'])
t8.to_csv(r'temps/t8.csv', index = False, header=True)

t9 = compute_themaScore(dataset[240000:270000])
t9 = t9.drop(columns=['genres'])
t9.to_csv(r'temps/t9.csv', index = False, header=True)

t10 = compute_themaScore(dataset[270000:290000])
t10 = t10.drop(columns=['genres'])
t10.to_csv(r'temps/t10.csv', index = False, header=True)
'''


t11 = compute_themaScore(dataset[290000:])
t11 = t11.drop(columns=['genres'])
t11.to_csv(r'temps/t11.csv', index = False, header=True)



#dataset.to_csv(r'export/dataset.csv', index = False, header=True)
#dataset = dataset.drop(columns=['genres'])
