==> 05. Imdb Score Model/03.Ter Cleaning Dataset.ipynb <==
import pandas as pd
import numpy as np

dataset = pd.read_csv(f'temps/datasetWithPonderation.csv', delimiter=',', on_bad_lines='skip')
print(len(dataset))
'''
for i in range(2,12):
    importDF = pd.read_csv(f'temps/t{i}.csv', delimiter=',', on_bad_lines='skip')
    dataset = dataset.append(importDF)
    print(len(dataset))
'''
dataset = dataset.rename(columns={"originalTitle": "title", "startYear": "year", "primaryName" : "director", "parental_advisor" : "parentalAdvisor", "averageRating" : "imdbRating"})
dataset = dataset.drop(columns=['title', 'director', 'resume'])
dataset = dataset[dataset['genre'] != np.nan ]
#dataset = dataset.dropna(subset=['year'])
#dataset = dataset.dropna(subset=['imdbRating'])
dataset = dataset.dropna(subset=['parentalAdvisor'])
dataset = dataset.dropna(subset=['duree'])
#dataset = dataset.dropna(subset=['genre'])
dataset = dataset[dataset['themaScore'] != 0 ]
dataset = dataset.reindex()
print(len(dataset))
#print(dataset.columns)

dataset = dataset.sort_values(by=['themaScore'], ascending=False)
dataset = dataset.drop_duplicates(subset=['tconst'])

print(len(dataset))
display(dataset.head())

target_name = 'imdbRating'
print(len(dataset))
# ReIndex the dataset for put the TARGET on last column. 
dataset = dataset.reindex(columns = [col for col in dataset.columns if col != target_name] + [target_name])

print('Taille Dataset :')
print(len(dataset)) 
print('Valeurs manquantes en % :')
100*dataset.isnull().sum()/dataset.shape[0]

dataset.to_csv(r'export/dataset.csv', index = False, header=True)
